trustnet-backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
Pillow==10.4.0
python-decouple==3.8
django-filter==24.3
dj-database-url==2.1.0
requests==2.32.3
